feat(tests): add IpAddress model tests and rename SaveRequestMiddlewareTest namespace
